Move model loading function to training utils
